+++ Libs/libv3d/libv3d.h
@@ -6,7 +6,7 @@
 /*   By: jubarbie <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/08 11:51:03 by jubarbie          #+#    #+#             */
-/*   Updated: 2016/12/03 16:40:37 by jubarbie         ###   ########.fr       */
+/*   Updated: 2016/12/03 16:58:01 by jubarbie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,20 +19,20 @@
 
 typedef struct	s_v3d
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }				t_v3d;
 
-t_v3d			v3d(float x, float y, float z);
+t_v3d			v3d(double x, double y, double z);
 void			print_v3d(t_v3d v);
 t_v3d			add_v3d(t_v3d v1, t_v3d v2);
 t_v3d			sub_v3d(t_v3d v1, t_v3d v2);
-t_v3d			smul_v3d(t_v3d v, float	nb);
-float			dot_v3d(t_v3d v1, t_v3d v2);
+t_v3d			smul_v3d(t_v3d v, double nb);
+double			dot_v3d(t_v3d v1, t_v3d v2);
 t_v3d			cross_v3d(t_v3d v1, t_v3d v2);
 t_v3d			unit_v3d(t_v3d v);
-float			length_v3d(t_v3d v);
-float			cos_v3d(t_v3d v1, t_v3d v2);
+double			length_v3d(t_v3d v);
+double			cos_v3d(t_v3d v1, t_v3d v2);
 
 #endif
